fix: link get clicked when panning graph
do not initiate link click in following scenario:
graph is panned by downing left mouse button on link, moving mouse while holding left button, and then upping mouse.

mouse up in the end - is just end of graph panning, not click on the link. and not both as in this bug.